fix(access): resolve eslint react-refresh and unused var errors
Extract useChallengeRowColumns hook, ResourceLink component, and
challenge helpers into separate files to satisfy react-refresh
only-export-components rule. Remove unused member destructure
and AuthzChallenge import.
