+++ client/dashboard/src/pages/access/ChallengesTab.tsx
@@ -1,4 +1,3 @@
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import {
   Select,
   SelectContent,
@@ -7,41 +6,19 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { SkeletonTable } from "@/components/ui/skeleton";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { Type } from "@/components/ui/type";
-import { HumanizeDateTime } from "@/lib/dates";
 import { cn } from "@/lib/utils";
-import { useSession } from "@/contexts/Auth";
-import { useSlugs } from "@/contexts/Sdk";
 import type { AuthzChallenge } from "@gram/client/models/components/authzchallenge.js";
 import { useChallenges } from "@gram/client/react-query/challenges.js";
-import { useMembers } from "@gram/client/react-query/members.js";
-import {
-  Badge as MoonshineBadge,
-  Column,
-  Table,
-} from "@speakeasy-api/moonshine";
-import {
-  Building2,
-  Check,
-  ChevronRight,
-  FolderOpen,
-  KeyRound,
-  Plug,
-} from "lucide-react";
+import { Badge as MoonshineBadge, Table } from "@speakeasy-api/moonshine";
+import { Check } from "lucide-react";
 import { useMemo, useState } from "react";
-import { Link, useSearchParams } from "react-router";
+import { useSearchParams } from "react-router";
+import { useChallengeRowColumns } from "./useChallengeRowColumns";
 import { useGrantFlow } from "./useGrantFlow";
 
 export type { AuthzChallenge } from "@gram/client/models/components/authzchallenge.js";
-export { invalidateAllChallenges } from "@gram/client/react-query/challenges.js";
 
 type Outcome = AuthzChallenge["outcome"];
-type Reason = AuthzChallenge["reason"];
 type OutcomeFilter = "all" | "deny" | "allow" | "resolved";
 
 export function OutcomeBadge({
@@ -73,71 +50,6 @@ export function OutcomeBadge({
     <MoonshineBadge variant={c.variant}>
       <MoonshineBadge.Text>{c.label}</MoonshineBadge.Text>
     </MoonshineBadge>
-  );
-}
-
-export function getInitials(identifier: string): string {
-  const name = identifier.split("@")[0] ?? identifier;
-  return name.slice(0, 2).toUpperCase();
-}
-
-function ResourceLink({
-  challenge,
-  orgSlug,
-  projectMap,
-}: {
-  challenge: AuthzChallenge;
-  orgSlug: string;
-  projectMap: Map<string, { slug: string; name: string }>;
-}) {
-  const { resourceKind, resourceId, projectId } = challenge;
-
-  if (!resourceKind || !resourceId) {
-    return (
-      <Type variant="body" className="text-muted-foreground text-sm">
-        —
-      </Type>
-    );
-  }
-
-  let to: string | null = null;
-  let label = resourceId;
-  let IconEl: typeof Building2 | null = null;
-
-  if (resourceKind === "org") {
-    label = "Organization";
-    IconEl = Building2;
-    to = `/${orgSlug}/settings`;
-  } else if (resourceKind === "project") {
-    const proj = projectMap.get(resourceId);
-    label = proj?.name ?? resourceId;
-    IconEl = FolderOpen;
-    to = proj ? `/${orgSlug}/projects/${proj.slug}` : null;
-  } else if (resourceKind === "mcp") {
-    label = resourceId;
-    IconEl = Plug;
-    const proj = projectId ? projectMap.get(projectId) : undefined;
-    to = proj ? `/${orgSlug}/projects/${proj.slug}/mcp/${resourceId}` : null;
-  }
-
-  if (to) {
-    return (
-      <Link
-        to={to}
-        className="inline-flex items-center gap-1.5 truncate text-sm text-blue-600 underline underline-offset-4 hover:text-blue-500 dark:text-blue-400 dark:hover:text-blue-300"
-      >
-        {IconEl && <IconEl className="h-3.5 w-3.5 shrink-0 opacity-60" />}
-        <span className="truncate">{label}</span>
-        <ChevronRight className="h-3 w-3 shrink-0" />
-      </Link>
-    );
-  }
-
-  return (
-    <span className="text-muted-foreground inline-flex items-center gap-1.5 truncate text-sm">
-      {IconEl && <IconEl className="h-3.5 w-3.5 shrink-0" />}
-      <span className="truncate">{label}</span>
-    </span>
   );
 }
 
@@ -173,217 +85,6 @@ function FilterPill({
         {count}
       </span>
     </button>
-  );
-}
-
-function reasonLabel(reason: Reason): string {
-  switch (reason) {
-    case "grant_matched":
-      return "Access granted — a matching role permission was found.";
-    case "no_grants":
-      return "No permissions configured for this identity.";
-    case "scope_unsatisfied":
-      return "The identity's roles don't include this permission.";
-    case "invalid_check":
-      return "The authorization check was malformed or invalid.";
-    case "rbac_skipped_apikey":
-      return "API keys bypass role checks — access was allowed directly.";
-    case "dev_override":
-      return "Access allowed via a development override.";
-  }
-}
-
-export function useChallengeRowColumns(
-  animatingOutIds?: Set<string>,
-  groupCounts?: Map<string, number>,
-): Column<AuthzChallenge>[] {
-  const { orgSlug } = useSlugs();
-  const { organization } = useSession();
-  const { data: membersData } = useMembers();
-  const projectMap = useMemo(() => {
-    const m = new Map<string, { slug: string; name: string }>();
-    for (const p of organization.projects) {
-      m.set(p.id, { slug: p.slug, name: p.name });
-    }
-    return m;
-  }, [organization.projects]);
-  const memberMap = useMemo(() => {
-    const m = new Map<string, { email: string; photoUrl?: string }>();
-    for (const member of membersData?.members ?? []) {
-      m.set(member.id, { email: member.email, photoUrl: member.photoUrl });
-    }
-    return m;
-  }, [membersData]);
-
-  const rowFade = (row: AuthzChallenge) =>
-    animatingOutIds?.has(row.id)
-      ? "opacity-0 transition-opacity duration-1000"
-      : (row.outcome === "allow" || row.resolvedAt) &&
-        "opacity-40 transition-opacity duration-1000";
-
-  return useMemo(
-    () => [
-      {
-        key: "avatar",
-        header: "",
-        width: "44px",
-        render: (row: AuthzChallenge) => {
-          const isApiKey = row.principalType === "api_key";
-          const display = row.userEmail ?? row.principalUrn;
-          return (
-            <div className={cn(rowFade(row))}>
-              <Avatar className="h-8 w-8">
-                {row.photoUrl && (
-                  <AvatarImage src={row.photoUrl} alt={display} />
-                )}
-                <AvatarFallback className="text-[11px]">
-                  {isApiKey ? (
-                    <KeyRound className="h-4 w-4" />
-                  ) : (
-                    getInitials(display)
-                  )}
-                </AvatarFallback>
-              </Avatar>
-            </div>
-          );
-        },
-      },
-      {
-        key: "identity",
-        header: "Identity",
-        width: "1fr",
-        render: (row: AuthzChallenge) => (
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Type
-                variant="body"
-                className={cn("truncate text-sm font-medium", rowFade(row))}
-              >
-                {row.userEmail ?? row.principalUrn}
-              </Type>
-            </TooltipTrigger>
-            {row.roleSlugs.length > 0 && (
-              <TooltipContent side="bottom">
-                Roles: {row.roleSlugs.join(", ")}
-              </TooltipContent>
-            )}
-          </Tooltip>
-        ),
-      },
-      {
-        key: "outcome",
-        header: "Outcome",
-        width: "85px",
-        render: (row: AuthzChallenge) => (
-          <div className={cn(rowFade(row))}>
-            <OutcomeBadge outcome={row.outcome} resolved={!!row.resolvedAt} />
-          </div>
-        ),
-      },
-      {
-        key: "scope",
-        header: "Required Scope",
-        width: "1fr",
-        render: (row: AuthzChallenge) => (
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <code
-                className={cn(
-                  "bg-muted rounded px-1.5 py-0.5 font-mono text-xs",
-                  rowFade(row),
-                )}
-              >
-                {row.scope}
-              </code>
-            </TooltipTrigger>
-            <TooltipContent side="bottom" className="max-w-xs">
-              <p className="text-xs">
-                {reasonLabel(row.reason)}
-                {row.evaluatedGrantCount > 0 &&
-                  ` (${row.matchedGrantCount} of ${row.evaluatedGrantCount} grants matched)`}
-              </p>
-            </TooltipContent>
-          </Tooltip>
-        ),
-      },
-      {
-        key: "resource",
-        header: "Resource",
-        width: "1fr",
-        render: (row: AuthzChallenge) => (
-          <div className={cn("min-w-0 overflow-hidden", rowFade(row))}>
-            <ResourceLink
-              challenge={row}
-              orgSlug={orgSlug ?? ""}
-              projectMap={projectMap}
-            />
-          </div>
-        ),
-      },
-      {
-        key: "resolvedBy",
-        header: "Resolved By",
-        width: "90px",
-        render: (row: AuthzChallenge) => {
-          if (!row.resolvedBy) {
-            return (
-              <Type variant="body" className="text-muted-foreground/40 text-sm">
-                —
-              </Type>
-            );
-          }
-          const userId = row.resolvedBy.replace(/^user:/, "");
-          const member = memberMap.get(userId);
-          const display = member?.email ?? row.resolvedBy;
-          return (
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Avatar className="h-7 w-7">
-                  {member?.photoUrl && (
-                    <AvatarImage src={member.photoUrl} alt={display} />
-                  )}
-                  <AvatarFallback className="text-[10px]">
-                    {getInitials(display)}
-                  </AvatarFallback>
-                </Avatar>
-              </TooltipTrigger>
-              <TooltipContent>{display}</TooltipContent>
-            </Tooltip>
-          );
-        },
-      },
-      {
-        key: "timestamp",
-        header: "Time",
-        width: "160px",
-        render: (row: AuthzChallenge) => {
-          const count = groupCounts?.get(row.id) ?? 1;
-          return (
-            <div className={cn("flex items-center gap-1.5", rowFade(row))}>
-              <Tooltip delayDuration={500}>
-                <TooltipTrigger asChild>
-                  <Type
-                    variant="body"
-                    className="text-muted-foreground cursor-default text-sm whitespace-nowrap underline decoration-dotted underline-offset-4"
-                  >
-                    <HumanizeDateTime date={row.timestamp} />
-                  </Type>
-                </TooltipTrigger>
-                <TooltipContent>
-                  {row.timestamp.toLocaleString()}
-                </TooltipContent>
-              </Tooltip>
-              {count > 1 && (
-                <span className="text-muted-foreground bg-muted rounded-full px-1.5 py-0.5 text-[10px] font-medium tabular-nums">
-                  ×{count}
-                </span>
-              )}
-            </div>
-          );
-        },
-      },
-    ],
-    [orgSlug, projectMap, memberMap, animatingOutIds, groupCounts],
   );
 }
